use size_t

+++ Core/include/Acts/Clusterization/Clusterization.hpp
@@ -97,8 +97,8 @@ struct DefaultConnect<Cell, 2> : public Connect2D<Cell> {
 template <typename CellCollection, std::size_t GridDim = 2,
           typename Connect =
               DefaultConnect<typename CellCollection::value_type, GridDim>>
-std::vector<std::uint8_t> labelClusters(CellCollection& cells,
-                                        Connect&& connect = Connect());
+std::vector<std::size_t> labelClusters(CellCollection& cells,
+                                       Connect&& connect = Connect());
 
 /// @brief mergeClusters
 ///
